[Refactor] Myinfo 코드 분리
